Extract Acls test logic from AclsDummySpec.
